letter_project/requirements.txt
django
gunicorn
psycopg2-binary
djangorestframework
dj-database-url
whitenoise
django-cors-headers